Move shared runners code into internal/runbits.
Also implemented other PR feedback.
